backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
redis>=5.2.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
alembic>=1.14.0
python-multipart>=0.0.20
aiofiles>=24.1.0
scikit-learn>=1.3.0
ollama>=0.1.0
python-dateutil>=2.8.0
imagehash>=4.3.1

# Image-to-Video AI (Wan 2.1)
diffusers>=0.31.0
transformers>=4.46.0
torch>=2.5.0
accelerate>=1.2.0
pillow>=11.0.0
imageio>=2.36.0
imageio-ffmpeg>=0.5.1
sentencepiece>=0.2.0

# Video Merging with RIFE
tqdm>=4.35.0
sk-video>=1.1.10
opencv-python>=4.1.2
moviepy>=2.0.0
